shopping site update v1.7

+++ src/main/java/com/example/boardproject/controller/SearchController.java
@@ -18,7 +18,7 @@ public class SearchController {
     }
 
     @GetMapping("/product/{pId}")
-    String showContents4(@PathVariable Long pId, Model model) {
+    String showContents4(@PathVariable int pId, Model model) {
         model.addAttribute("allContents", searchRepository.searchList(pId));
         model.addAttribute("pId", pId);
         return "list";
